% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bin.prob.B.R
\name{calc.bin.prob.B}
\alias{calc.bin.prob.B}
\title{This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.}
\usage{
calc.bin.prob.B(n.vec, p.vec)
}
\arguments{
\item{n.vec}{vector of number of trials}

\item{p.vec}{vector of probabilities}
}
\value{
vector of binary probability, dichotomization threshold
}
\description{
This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.
}
