% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bin.prob.NB.R
\name{calc.bin.prob.NB}
\alias{calc.bin.prob.NB}
\title{This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.}
\usage{
calc.bin.prob.NB(r.vec, prob.vec)
}
\arguments{
\item{r.vec}{vector of number of trials}

\item{prob.vec}{vector of probabilities}
}
\value{
vector of binary probability, dichotomous threshold
}
\description{
This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.
}
