% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_cont.R
\name{discrete_cont}
\alias{discrete_cont}
\title{Check positive definiteness of the intermediate matrix}
\usage{
discrete_cont(
  marginal,
  Sigma,
  support = list(),
  Spearman = FALSE,
  epsilon = 1e-06,
  maxit = 100
)
}
\arguments{
\item{marginal}{a list of \eqn{k} elements, where \eqn{k} is the number of variables.
The \eqn{i}-th element of \code{marginal} is the vector of the cumulative probabilities defining the marginal distribution of the \eqn{i}-th component of the  multivariate variable. If the \eqn{i}-th component can take \eqn{k_i} values, the \eqn{i}-th element of \code{marginal} will contain \eqn{k_i-1} probabilities (the \eqn{k_i}-th is obviously 1 and shall not be included).}

\item{Sigma}{the target correlation matrix of the discrete variables}

\item{support}{a list of \eqn{k} elements, where \eqn{k} is the number of variables. The \eqn{i}-th element of \code{support} is the vector containing the ordered values of the support of the \eqn{i}-th variable. By default, the support of the \eqn{i}-th variable is \eqn{1,2,...,k_i}}

\item{Spearman}{A logical flag indicating whether Spearman correlation should be used}

\item{epsilon}{tolerance of the algorithm convergence}

\item{maxit}{maximum iterations of the algorithm to correct PD matrix}
}
\value{
No return values; called it to check parameter inputs
}
\description{
Check positive definiteness of the intermediate matrix
}
