% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.binaryVar.R
\name{generate.binaryVar}
\alias{generate.binaryVar}
\title{Generate multivariate Binary data using the Emrich and Piedmonte (1991)
Approach}
\usage{
generate.binaryVar(nObs, prop.vec.bin, corr.mat)
}
\arguments{
\item{nObs}{number of observations}

\item{prop.vec.bin}{probability of binary variables in a vector}

\item{corr.mat}{Correlation matrix}
}
\value{
multivariate Binary Data
}
\description{
Generate multivariate Binary data using the Emrich and Piedmonte (1991)
Approach
}
