% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genGPD.R
\name{genGPD}
\alias{genGPD}
\title{This function generates multivariate GPD data}
\usage{
genGPD(no.rows, binObj)
}
\arguments{
\item{no.rows}{number of data}

\item{binObj}{intermediate correlation matrix object}
}
\value{
generated GPD data with user's specifications
}
\description{
This function generates multivariate GPD data
}
