% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMix.R
\name{genMix}
\alias{genMix}
\title{This function generates multivariate Mixed Discrete Data}
\usage{
genMix(no.rows, binObj)
}
\arguments{
\item{no.rows}{number of data}

\item{binObj}{intermediate correlation matrix object}
}
\value{
generated mixed discrete data with user's specifications
}
\description{
This function generates multivariate Mixed Discrete Data
}
