% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genNB.R
\name{genNB}
\alias{genNB}
\title{This function generates multivariate NB data}
\usage{
genNB(no.rows, binObj)
}
\arguments{
\item{no.rows}{number of data}

\item{binObj}{intermediate correlation matrix object}
}
\value{
generated NB data with user's specifications
}
\description{
This function generates multivariate NB data
}
