% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordcont_nearPD.R
\name{ordcont_nearPD}
\alias{ordcont_nearPD}
\title{Check positive definiteness of the intermediate matrix}
\usage{
ordcont_nearPD(
  marginal,
  Sigma,
  support = list(),
  Spearman = FALSE,
  epsilon = 1e-06,
  maxit = 100
)
}
\arguments{
\item{marginal}{Marginal vector}

\item{Sigma}{Covariance matrix}
}
\value{
No return values; called it to check parameter inputs
}
\description{
Check positive definiteness of the intermediate matrix
}
