% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.GPDparameters.R
\name{validation.GPDparameters}
\alias{validation.GPDparameters}
\title{Validate if the input GPD parameters are within feasible range}
\usage{
validation.GPDparameters(theta.vec, lambda.vec)
}
\arguments{
\item{theta.vec}{Vector of theta values}

\item{lambda.vec}{Vector of lambda values}
}
\value{
No return values; called it to check parameter inputs
}
\description{
This function returns the sum of two numbers.
}
\examples{
validation.GPDparameters(theta.vec = c(3, 2), lambda.vec = c(0.4, 0.2))
}
