% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinToB.R
\name{BinToB}
\alias{BinToB}
\title{Convert Multivariate Binary Data Back to the Original Binomial Scale}
\usage{
BinToB(prop.vec.bin, BProp, Mlocation, bin.data)
}
\arguments{
\item{prop.vec.bin}{A vector of binary probabilities}

\item{BProp}{Binary proportion}

\item{Mlocation}{Indices of the medians in the vector}

\item{bin.data}{Generated multivariate binary data.}
}
\value{
A list containing the multivariate binomial data and its correlation matrix
}
\description{
This function implements step 5 of the algorithm described in the paper.
It converts multivariate binary data back to the original binomial scale.
}
