% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinToGPD.R
\name{BinToGPD}
\alias{BinToGPD}
\title{This function implements Step 5 of the algorithm
It converts the multivariate binary data back to the original GPD scale}
\usage{
BinToGPD(prop.vec.bin, GPDprop, Mlocation, bin.data)
}
\arguments{
\item{prop.vec.bin}{vector of binary probabilities}

\item{GPDprop}{GPD proportion}

\item{Mlocation}{locations of median in the vector}

\item{bin.data}{generated multivariate binary data}
}
\value{
multivariate GPD data and its correlation matrix
}
\description{
This function implements Step 5 of the algorithm
It converts the multivariate binary data back to the original GPD scale
}
