% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinToGPD.R
\name{BinToGPD}
\alias{BinToGPD}
\title{Convert multivariate binary data back to the original generalized poisson scale}
\usage{
BinToGPD(prop.vec.bin, GPDprop, Mlocation, bin.data)
}
\arguments{
\item{prop.vec.bin}{A vector of binary probabilities.}

\item{GPDprop}{Generalized Poisson distribution proportions}

\item{Mlocation}{Indices of the medians in the vector}

\item{bin.data}{Generated multivariate binary data}
}
\value{
A list containing the multivariate GPD data and its correlation matrix
}
\description{
This function implements Step 5 of the algorithm.
It converts multivariate binary data back to the original GPD scale.
}
