% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinToNB.R
\name{BinToNB}
\alias{BinToNB}
\title{This function implements Step 5 of the algorithm.
It converts the multivariate binary data back to the original GPD scale}
\usage{
BinToNB(prop.vec.bin, NBprop, Mlocation, bin.data)
}
\arguments{
\item{prop.vec.bin}{vector of binary probabilities}

\item{NBprop}{NB proportion}

\item{Mlocation}{locations of median in the vector}

\item{bin.data}{generated multivariate binary data}
}
\value{
multivariate NB data and its correlation matrix
}
\description{
This function implements Step 5 of the algorithm.
It converts the multivariate binary data back to the original GPD scale
}
