% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bin.prob.GPD.R
\name{calc.bin.prob.GPD}
\alias{calc.bin.prob.GPD}
\title{This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.}
\usage{
calc.bin.prob.GPD(theta.vec, lambda.vec)
}
\arguments{
\item{theta.vec}{vector of theta values}

\item{lambda.vec}{vector of lambda values}
}
\value{
vector of binary probability, dichotomous threshold
}
\description{
This function implements Step 1 of the algorithm.
It collapses the discrete outcome to binary ones for each variable.
}
