% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genB.R
\name{genB}
\alias{genB}
\title{Generates multivariate binomial data from binary parameters}
\usage{
genB(no.rows, binObj)
}
\arguments{
\item{no.rows}{number of data}

\item{binObj}{intermediate correlation matrix object}
}
\value{
generated Binomial data with user's specifications
}
\description{
Generates multivariate binomial data from binary parameters
}
