% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genB.R
\name{genB}
\alias{genB}
\title{This function generates multivariate Binomial data}
\usage{
genB(no.rows, binObj)
}
\arguments{
\item{no.rows}{number of data}

\item{binObj}{intermediate correlation matrix object}
}
\value{
generated Binomial data with user's specifications
}
\description{
This function generates multivariate Binomial data
}
