% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGpoisPMF.R
\name{GetGpoisPMF}
\alias{GetGpoisPMF}
\title{Get PMF of Generalized Poisson distribution}
\usage{
GetGpoisPMF(p, theta, lambda, details = FALSE)
}
\arguments{
\item{p}{probability of GDP poisson pmf}

\item{theta}{GPD theta value}

\item{lambda}{GPD lambda value}
}
\value{
a PMF table
}
\description{
This function returns PMF of GPD
}
