% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGpoisPMF.R
\name{GetGpoisPMF}
\alias{GetGpoisPMF}
\title{Get probability mass function of Generalized Poisson distribution}
\usage{
GetGpoisPMF(p, theta, lambda, details = FALSE)
}
\arguments{
\item{p}{probability of generalized Poisson distribution}

\item{theta}{GPD theta value}

\item{lambda}{GPD lambda value}

\item{details}{A logical flag indicating computation information should be returned}
}
\value{
a PMF table
}
\description{
This function returns a table of the probability mass function of GPD
}
