% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGpoisPMF.R
\name{GetGpoisPMF}
\alias{GetGpoisPMF}
\title{Get probability mass function of generalized Poisson distribution}
\usage{
GetGpoisPMF(p, theta, lambda, details = FALSE)
}
\arguments{
\item{p}{Probability parameter for the generalized Poisson distribution.}

\item{theta}{Theta parameter(s) for the generalized Poisson distribution.}

\item{lambda}{Lambda parameter(s) for the generalized Poisson distribution.}

\item{details}{Logical; if TRUE, additional computation information is returned.}
}
\value{
A table containing the probability mass function.
}
\description{
This function returns a table of the probability mass function of generalized Poisson distribution
}
