% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileGpois.R
\name{QuantileGpois}
\alias{QuantileGpois}
\title{This function computes the quantile of Generalized Poisson}
\usage{
QuantileGpois(p, theta, lambda, details = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{theta}{vector of theta}

\item{lambda}{vector of lambda}

\item{details}{A logical flag to return the computational details}
}
\value{
the quantile of GPD
}
\description{
This function computes the quantile of Generalized Poisson
}
