% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinaryCorr.B.R
\name{simBinaryCorr.B}
\alias{simBinaryCorr.B}
\title{This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.}
\usage{
simBinaryCorr.B(n.vec, p.vec, CorrMat, no.rows, steps = 0.025)
}
\arguments{
\item{n.vec}{vector of number of trials}

\item{p.vec}{vector of probabilities}

\item{CorrMat}{specified Correlation matrix}

\item{no.rows}{number of observations for generating Multivariate Binary data}

\item{steps}{Fraction of difference between the current and target matrix to be added in each iteration.}
}
\value{
intermediate multivariate binary Correlation matrix
}
\description{
This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.
}
