% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinaryCorr.GPD.R
\name{simBinaryCorr.GPD}
\alias{simBinaryCorr.GPD}
\title{This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.}
\usage{
simBinaryCorr.GPD(theta.vec, lambda.vec, CorrMat, no.rows, steps = 0.025)
}
\arguments{
\item{theta.vec}{vector of theta values}

\item{lambda.vec}{vector of lambda values}

\item{CorrMat}{specified Correlation matrix}

\item{no.rows}{number of observations for generating Multivariate Binary data}
}
\value{
intermediate multivariate binary Correlation matrix
}
\description{
This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.
}
