% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBinaryCorr.NB.R
\name{simBinaryCorr.NB}
\alias{simBinaryCorr.NB}
\title{This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.}
\usage{
simBinaryCorr.NB(r.vec, prob.vec, CorrMat, no.rows, steps = 0.025)
}
\arguments{
\item{r.vec}{vector of number of trials}

\item{prob.vec}{vector of probabilities}

\item{CorrMat}{specified Correlation matrix}

\item{no.rows}{number of observations for generating Multivariate Binary data}
}
\value{
intermediate multivariate binary Correlation matrix
}
\description{
This function implements Step 2 of the algorithm
It calculates the intermediate binary correlations.
}
