% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.Bparameters.R
\name{validation.Bparameters}
\alias{validation.Bparameters}
\title{Validate if the input Binomial parameters are within feasible range}
\usage{
validation.Bparameters(B.n.vec, B.prob.vec)
}
\arguments{
\item{B.n.vec}{Vector of number of trials}

\item{B.prob.vec}{Vector of probability}
}
\value{
No return values; called it to check parameter inputs
}
\description{
This function returns the sum of two numbers.
}
\examples{
validation.Bparameters(B.n.vec = c(10, 15), B.prob.vec = c(0.4, 0.2))
}
