% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.NBparameters.R
\name{validation.NBparameters}
\alias{validation.NBparameters}
\title{Validate if the input NB parameters are within feasible range}
\usage{
validation.NBparameters(NB.r.vec, NB.prob.vec)
}
\arguments{
\item{NB.r.vec}{Vector of number of trials parameters}

\item{NB.prob.vec}{Vector of probabilities}
}
\value{
No return values; called it to check parameter inputs
}
\description{
Validate if the input NB parameters are within feasible range
}
\examples{
validation.NBparameters(NB.r.vec = c(10, 15), NB.prob.vec = c(0.7, 0.5))
}
